Remove permalink from DB
We are able to reconstruct permalink from channel and message ts, so we
don't need to store it in DB.

As a part of this change, use direct deep-linking into slack as a
replacement for former permalink. This is better UX as the former just
redirected user to the slack deep link.