Fix #10, load all unprocessed messages on startup
When the bot is not running, new messages could appear in the watched
channels. These messages are left unprocessed since the bot is
triggered only on a new message event. This issue fixes this problem
by looping through all the channels the bot is a member of on every
startup. For each channel, the latest message stored in the database
determines which message was the last one processed, every newer
message is loaded to the database.