Fix the emoji in the message list header
The corresponding emoji of the selected tag is shown in the header
incorrectly. The corresponding emojis of these special tags are
prefixed with two underscores in Slack. This fact was not taken
into consideration in the code before. Also, for more pleasing
visuals, the emoji was moved to the h2 element.