Route the rest of the GET requests to the client
Since the client now has its routes defined, let's make sure the server
sends any unrecognized GET requests to the client to deal with it.
Since the path to the client's 'index.html' contained '..', the 'express'
framework had thrown forbidden errors in this new configuration.
To prevent this, the path used was joined together.