Load and serve users mentioned in replies
The previous implementation only loaded users mentioned in reports.
This caused the client to crash when a report replied to a
non-report message mentioning some user, because we display
the starting message of a thread, even if it's not a report.
The starting message is also included when querying replies, so
it's enough to search for mentions in replies.

Fixes #30.